First iteration of working prototype of ext server
Currently, each extension directory in this repository is its own main
package, compiled separately, built into its own Docker image, run in
its own container, with its own port, own internal loadbalancer, own
systemd service unit etc. This makes is cumbersome to add new
extensions.

This commit marks the first working prototype of building an extension
"server". The main package handles requests for all extensions. The
server will be a single process, in a single container, running on a
single port.
