--- Dockerfile.bmc_store_password
@@ -1,6 +1,0 @@
-FROM golang:1.19-alpine3.17
-ADD . /go/src/github.com/m-lab/epoxy-extensions
-WORKDIR /go/src/github.com/m-lab/epoxy-extensions
-RUN apk add --no-cache git
-RUN go install -v ./bmc_store_password
-ENTRYPOINT ["/go/bin/bmc_store_password"]
