--- Dockerfile.token_server
@@ -1,5 +1,0 @@
-FROM golang:1.19
-ADD . /go/src/github.com/m-lab/epoxy-extensions
-WORKDIR /go/src/github.com/m-lab/epoxy-extensions
-RUN go install -v ./token_server
-ENTRYPOINT ["/go/bin/token_server"]
