+++ allocate_k8s_token/allocate_k8s_token.go
@@ -1,26 +1,5 @@
-// Copyright 2016 k8s-support Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//////////////////////////////////////////////////////////////////////////////
-
-// The token_server implements the epoxy extension API and provides a way for
-// machines booting with epoxy to allocate a k8s token, necessary for joining
-// the cluster.
-//
-// To deploy the token_server, the ePoxy server must have an extension
-// registered that maps an operation name to this server, e.g.:
-//
-//	"allocate_k8s_token" -> "http://localhost:8800/allocate_k8s_token"
+// allocate_k8s_token implements the epoxy extension API and provides a way for
+// machines booting with epoxy to obtain a bootstrap token to join the cluster.
 package allocate_k8s_token
 
 import (
@@ -34,10 +13,12 @@ var (
 	localCommander commander = &runCommand{}
 )
 
+// commander is an interface that is used to wrap os/exec.Command() for testing purposes.
 type commander interface {
 	Command(prog string, args ...string) ([]byte, error)
 }
 
+// runCommand implements the commander interface.
 type runCommand struct{}
 
 // Command takes a program name and arguments as parameters and hands those off
@@ -54,18 +35,22 @@ type TokenGenerator interface {
 	Response(version string) ([]byte, error)
 }
 
+// k8sTokenGenerator implements the TokenGenerator interface.
 type k8sTokenGenerator struct {
 	Command       string
-	TokenResponse tokenResponse
+	TokenResponse TokenResponse
 }
 
-type tokenResponse struct {
+// TokenResponse represents data used in responses to allocate_k8s_token
+// extension requests. For v1, only Token will be populated/returned, and for v2
+// all fields should have values and will be returned as JSON.
+type TokenResponse struct {
 	APIAddress string `json:"api_address"`
 	Token      string `json:"token"`
 	CAHash     string `json:"ca_hash"`
 }
 
-// Token generates a new k8s token.
+// Create generates a new k8s token.
 func (g *k8sTokenGenerator) Create(target string) error {
 	args := []string{
 		"token", "create", "--ttl", "5m", "--print-join-command",
@@ -99,10 +84,11 @@ func (g *k8sTokenGenerator) Response(version string) ([]byte, error) {
 	return json.Marshal(g.TokenResponse)
 }
 
+// New returns a partially populated k8sTokenGenerator
 func New(bindir string) *k8sTokenGenerator {
 	return &k8sTokenGenerator{
 		Command:       bindir + "/kubeadm",
-		TokenResponse: tokenResponse{},
+		TokenResponse: TokenResponse{},
 	}
 
 }
