+++ bmc_store_password/bmc_store_password.go
@@ -1,25 +1,5 @@
-// Copyright 2020 M-Lab Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//////////////////////////////////////////////////////////////////////////////
-
 // bmc_store_password implements the epoxy extension API and provides a way for
 // machines booting with epoxy to store the configured BMC password to GCD.
-//
-// To deploy bmc_store_password, the ePoxy server must have an extension
-// registered that maps an operation name to this server, e.g.:
-//
-//	"bmc_store_password" -> "http://localhost:8801/bmc_store_password"
 package bmc_store_password
 
 import (
