+++ server.go
@@ -1,22 +1,6 @@
-// Copyright 2023 M-Lab Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//////////////////////////////////////////////////////////////////////////////
-
 package main
 
 import (
-	"context"
 	"flag"
 	"fmt"
 	"log"
@@ -32,8 +16,6 @@ import (
 var (
 	fBinDir        string
 	fListenAddress string
-
-	mainCtx, mainCancel = context.WithCancel(context.Background())
 )
 
 // rootHandler implements the simplest possible handler for root requests,
@@ -52,8 +34,6 @@ func init() {
 }
 
 func main() {
-	defer mainCancel()
-
 	flag.Parse()
 
 	k8sToken := allocate_k8s_token.New(fBinDir)
